requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
audioread=2.1.4=py36_2
bleach=1.5.0=py36_0
bzip2=1.0.6=1
ca-certificates=2018.03.07=0
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h9745a5d_0
cudatoolkit=8.0=3
cudnn=7.0.5=cuda8.0_0
cycler=0.10.0=py36_0
dbus=1.10.22=0
decorator=4.2.1=py36_0
expat=2.2.5=0
ffmpeg=2.8.6=5
fontconfig=2.12.6=0
freetype=2.8.1=0
gettext=0.19.8.1=0
glib=2.55.0=0
gst-plugins-base=1.8.0=0
gstreamer=1.8.0=1
h5py=2.7.1=py36_2
hdf5=1.10.1=2
html5lib=0.9999999=py36_0
icu=58.2=0
intel-openmp=2018.0.0=8
joblib=0.11=py36_0
jpeg=9b=2
keras=2.1.5=py36_0
kiwisolver=1.0.1=py36_1
libffi=3.2.1=3
libgcc-ng=7.2.0=hdf63c60_3
libgfortran=3.0.0=1
libgfortran-ng=7.2.0=hdf63c60_3
libgpuarray=0.7.5=0
libiconv=1.15=0
libpng=1.6.34=0
libprotobuf=3.5.1=h6f1eeef_0
librosa=0.6.0=py36_0
libstdcxx-ng=7.2.0=hdf63c60_3
libxcb=1.13=0
libxml2=2.9.8=0
llvmlite=0.21.0=py36_0
mako=1.0.7=py36_0
markdown=2.6.11=py36_0
markupsafe=1.0=py36_0
matplotlib=2.2.2=py36_1
mkl=2018.0.2=1
mkl_fft=1.0.1=py36_1
mkl_random=1.0.1=py36_0
ncurses=5.9=10
numba=0.36.2=np114py36hc6662d5_0
numpy=1.14.2=py36hdbf6ddf_1
openssl=1.0.2o=h20670df_0
pandas=0.22.0=py36hf484d3e_0
pcre=8.41=1
pip=9.0.3=py36_0
protobuf=3.5.1=py36hf484d3e_0
pycparser=2.18=py36hf9f622e_1
pygpu=0.7.5=py36_0
pyparsing=2.2.0=py36_0
pyqt=5.6.0=py36_4
python=3.6.5=0
python-dateutil=2.7.2=py_0
pytz=2018.3=py_0
pyyaml=3.12=py36_1
qt=5.6.2=7
readline=7.0=0
resampy=0.2.0=py36_1
scikit-learn=0.19.1=py36h7aa7ec6_0
scipy=1.0.1=py36hfc37229_0
setuptools=39.0.1=py36_0
sip=4.18=py36_1
six=1.11.0=py36_1
sqlite=3.20.1=2
tensorflow-gpu=1.4.1=0
tensorflow-gpu-base=1.4.1=py36h01caf0a_0
tensorflow-tensorboard=1.5.1=py36hf484d3e_0
theano=1.0.1=py36_1
tk=8.6.7=0
tornado=5.0.1=py36_1
tqdm=4.19.9=py_0
werkzeug=0.14.1=py36_0
wheel=0.30.0=py36_2
x264=20131217=3
xorg-libxau=1.0.8=3
xorg-libxdmcp=1.1.2=3
xz=5.2.3=0
yaml=0.1.7=0
zlib=1.2.11=0
